Moved check to DeviceMonitorManager
