+++ src/NHM.DeviceMonitoring/packages.config
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="HidSharp" version="2.1.0" targetFramework="net48" />
+  <package id="LibreHardwareMonitorLib" version="0.8.8" targetFramework="net48" />
+  <package id="NvAPIWrapper.Net" version="0.8.1.101" targetFramework="net48" />
+  <package id="System.Management" version="4.7.0" targetFramework="net48" />
+</packages>
